Update GCC to 10.3.1, toolchain to Ubuntu 22.04, keyboard -O3 to -O2
This will not break compiling and flashing functionality with the
docker-compose setup, while ensuring we're running with some more
up-to-date version of GCC. Also, Ubuntu 16.04 is EOL (End-Of-Life) and
no longer supported, so we should not even use it anymore.

Finally, this change downgrades the optimization level for the keyboard
firmware(s) from -O3 to -O2. The reason being that newer versions of GCC
seem to produce bigger but different code for the keyboard firmwares,
resulting in unexpected keyboard matrix transmission behaviour. See the
respective Makefile for more information.